backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
transformers==4.10.0
torch==2.2.0
diffusers==0.25.0
python-dotenv==1.0.0
accelerate==0.27.0 